Updated Switch/Case
When you wish for two or more cases to have the same result, you can just have a case without a break and it will use the next case (if there is a break available). For example instead of doing it like

```
var variable = "matt";
switch(variable) {
    case "matt":
        console.log("matthew");
        break;
    case "matthew":
        console.log("matthew");
        break;
}
```

You can have it so it would be like 

```
var variable = "matt";
switch(variable) {
    case "matt":
    case "matthew":
        console.log("matthew");
        break;
}
```

Which will have the case move down to the next until there is a break to stop it. Note: Make sure you have it above the case you wish to fall on and not at random, otherwise it will cause problems.